화면에 점수 및 FPS 출력

+++ NetworkGamePractice/src/Graphics.h
@@ -29,7 +29,7 @@ public:
 	bool loadFont(const char* fontPath);
 
 	bool shouldClose();
-	void renderFrame(double elapsedTime);
+	void renderFrame(float elapsedTime);
 	void renderObject(GameObject targetObject);
 	void renderText(std::string text, float x, float y, float scale, glm::vec3 color);
